% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-actions.R
\docType{data}
\name{actions}
\alias{actions}
\title{Example trading actions}
\format{An object of class \code{tibble}.}
\source{
the close prices are retrieved from \href{https://kucoin.com}{KuCoin}
market's API.
}
\usage{
data(actions)
}
\description{
Example trading actions
}
\details{
This dataset provide trading actions data, which
created using \link{btcusdt} data and \link[=step_zigzag]{step_zigzag()}
for an example trading actions.
}
\examples{

# import library
library(quantrecipes)

# print the data
actions

}
\keyword{datasets}
