% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-btcusdt.R
\docType{data}
\name{btcusdt}
\alias{btcusdt}
\title{BTC/USDT historical data}
\format{An object of class \code{tibble}.}
\source{
\href{https://kucoin.com}{KuCoin} market's API.
}
\usage{
data(btcusdt)
}
\description{
BTC/USDT historical data
}
\details{
This historical data is obtained from \href{https://kucoin.com}{KuCoin}
market's API, which queried using
\href{https://github.com/bagasbgy/rucoin}{rucoin}. The dataset is ranging
from May 1, 2019 to July 31, 2019, and contains following informations:
\itemize{
\item \code{datetime}
\item \code{open}
\item \code{high}
\item \code{low}
\item \code{close}
\item \code{volume}
\item \code{turnover}
}
}
\examples{

# import library
library(quantrecipes)

# print the data
btcusdt

}
\keyword{datasets}
