% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_bbands.R
\name{step_bbands}
\alias{step_bbands}
\alias{tidy.step_bbands}
\title{Extract Bollinger Bands features}
\usage{
step_bbands(recipe, ..., ma_fun = TTR::SMA, n = 20, sd_mult = 2,
  prefix = "bbands", h = NULL, l = NULL, c = NULL,
  role = "predictor", trained = FALSE, skip = FALSE,
  id = rand_id("bbands"))

\method{tidy}{step_bbands}(x, ...)
}
\arguments{
\item{recipe}{A recipe object. The step will be added to the
sequence of operations for this recipe.}

\item{...}{Three (unquoted) column names, which, respectively,
representing the \code{"high"}, \code{"low"}, and \code{"close"} prices.}

\item{ma_fun}{A \code{function} to extract moving average elements, or a \code{character}
vector of length one which specify a moving average function.
Default to \link[TTR:SMA]{TTR::SMA}.}

\item{n}{A \code{numeric} vector of length one which specify
moving average window. The default is \code{20}.}

\item{sd_mult}{A \code{numeric} vector of length one which specify
standard deviation multiplier. The default is \code{2}.}

\item{prefix}{A \code{character} vector of length one that would be used
as a prefix to the created bollinger bands columns. Default to \code{"bbands"}.}

\item{h}{A container for the names of \code{"high"}. Leave to \code{NULL}
as it will populated by \code{\link[recipes:prep.recipe]{recipes::prep.recipe()}} function.}

\item{l}{A container for the names of \code{"low"}. Leave to \code{NULL}
as it will populated by \code{\link[recipes:prep.recipe]{recipes::prep.recipe()}} function.}

\item{c}{A container for the names of \code{"close"}. Leave to \code{NULL}
as it will populated by \code{\link[recipes:prep.recipe]{recipes::prep.recipe()}} function.}

\item{role}{For model terms created by this step, what analysis
role should they be assigned? By default, the function assumes
that the created bollinger bands columns will be used
as predictors in a model.}

\item{trained}{A logical to indicate if the necessary informations for
preprocessing have been estimated.}

\item{skip}{A logical. Should the step be skipped when the
recipe is baked by \code{\link[recipes:bake.recipe]{recipes::bake.recipe()}}? While all operations are baked
when \code{\link[recipes:prep.recipe]{recipes::prep.recipe()}} is run, some operations may not be able to be
conducted on new data (e.g. processing the outcome variable(s)).
Care should be taken when using \code{skip = TRUE} as it may affect
the computations for subsequent operations}

\item{id}{A character string that is unique to this step to identify it.}

\item{x}{A \code{step_bbands} object.}
}
\value{
An updated version of \code{recipe} with the new step
added to the sequence of existing steps (if any).
}
\description{
\code{step_bbands} creates a \strong{specification} of a recipe
step that will extract \strong{Bollinger Bands} features
from an asset price historical data.
}
\details{
The output from this step are several new column
which contains the extracted Bollinger Bands features:
\itemize{
\item \code{dn}
\item \code{ma}
\item \code{up}
\item \code{pctb}
\item \code{state}
\item \code{state_count}
\item \code{prev_state}
\item \code{prev_range}
\item \code{prev_break}
}
}
\examples{

# import libs
library(quantrecipes)

# an example recipes using built-in data
rec <- recipe(. ~ ., data = btcusdt) \%>\%
  step_bbands(high, low, close, ma_fun = TTR::SMA, n = 20, sd_mult = 2) \%>\%
  step_naomit(all_predictors()) \%>\%
  prep()

# get preprocessed data
juice(rec)

}
